+++ torchmetrics/functional/__init__.py
@@ -62,7 +62,7 @@ from torchmetrics.functional.regression.r2 import r2_score
 from torchmetrics.functional.regression.spearman import spearman_corrcoef
 from torchmetrics.functional.regression.symmetric_mape import symmetric_mean_absolute_percentage_error
 from torchmetrics.functional.regression.tweedie_deviance import tweedie_deviance_score
-from torchmetrics.functional.regression.wape import weighted_absolute_percentage_error
+from torchmetrics.functional.regression.wmape import weighted_mean_absolute_percentage_error
 from torchmetrics.functional.retrieval.average_precision import retrieval_average_precision
 from torchmetrics.functional.retrieval.fall_out import retrieval_fall_out
 from torchmetrics.functional.retrieval.hit_rate import retrieval_hit_rate
@@ -159,7 +159,7 @@ __all__ = [
     "translation_edit_rate",
     "universal_image_quality_index",
     "spectral_angle_mapper",
-    "weighted_absolute_percentage_error",
+    "weighted_mean_absolute_percentage_error",
     "word_error_rate",
     "word_information_lost",
     "word_information_preserved",